No longer expose the format string capability of the underlying library
Using it incorrectly or unintentionally can lead to JVM crashes or unintended
data being logged.
One example is a format string which contains %n to which a workaround was
applied in 0.2.5-1. Other occurrences of % can however cause problems too as they
will cause sd_journal_print/send to look for args which might not actually have
been supplied. For example (jprint :info "%x") will either log whatever value
is in the place a third arg would've been or crash the JVM if that place wasn't
allocated.